+++ Cargo.toml
@@ -8,6 +8,7 @@ members = ["examples",
            "hoomd-interaction",
            "hoomd-mc",
            "hoomd-microstate",
+           "hoomd-simulation",
            "hoomd-utility",
            "hoomd-vector",
            "tools/build-wasm-doc-examples",
@@ -23,6 +24,7 @@ default-members = ["examples",
                    "hoomd-interaction",
                    "hoomd-mc",
                    "hoomd-microstate",
+                   "hoomd-simulation",
                    "hoomd-utility",
                    "hoomd-vector",
                   ]
@@ -45,6 +47,7 @@ hoomd-gsd = { path = "hoomd-gsd" }
 hoomd-interaction = { path = "hoomd-interaction" }
 hoomd-mc = { path = "hoomd-mc" }
 hoomd-microstate = { path = "hoomd-microstate" }
+hoomd-simulation = { path = "hoomd-simulation" }
 hoomd-utility = { path = "hoomd-utility" }
 hoomd-vector = { path = "hoomd-vector" }
