+++ hoomd-simulation/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "hoomd-simulation"
+categories.workspace = true
+description = "Define simulation models in hoodm-rs."
+edition.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+license.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+
+[lints]
+workspace = true
+
+[dependencies]
+anyhow.workspace = true
